pyproject.toml
[tool.poetry]
name = "blade-logging"
version = "0.1.0"
description = "Blade Logging Technical Test"
authors = ["Elias Rhouzlane"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
dateutils = "^0.6.8"
python-dateutil = "^2.8.1"
tzlocal = "^2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
mypy = "^0.782"
taskipy = "^1.2.1"
pylint = "^2.5.3"
yapf = "^0.30.0"
faker = "^4.1.1"

[tool.taskipy.tasks]
lint_pylint = "pylint *.py"
lint_mypy = "mypy --config-file pyproject.toml tests main.py"
lint = "task lint_pylint && task lint_mypy"
test = "task lint_mypy && pytest"
dev = "task lint_mypy && python main.py"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[mypy]
python_version = 3.8
ignore_missing_imports = true

[pytest]
cache_dir = ".pytest_cache"
testpaths = "tests"

